requirements.txt
imageio >= 2.31.3
imageio-ffmpeg >= 0.4.9
mediapipe >= 0.10.3
PyAutoGUI >= 0.9.54
macmouse >= 0.7.4
nbstripout >= 0.6.0 # for removing notebook output from git
PySide6 >= 6.5.2
pyobjc-framework-Cocoa >= 9.2;platform_system == "Darwin"
qt-material >= 2.14
pygrabber >= 0.1.0
vidcontrol >= 0.1.1
PyYAML==6.0.1
numpy~=1.26.2
opencv-python~=4.8.1.78
sounddevice~=0.4.6
Pillow~=10.1.0